LayoutTests/canvas/philip/tests/2d.text.draw.baseline.ideographic.html

<!DOCTYPE html>
<title>Canvas test: 2d.text.draw.baseline.ideographic</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="2d.text.draw.baseline.alphabetic.html" title="2d.text.draw.baseline.alphabetic">
<link rel="next" href="2d.text.draw.baseline.hanging.html" title="2d.text.draw.baseline.hanging">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("../fonts/CanvasTest.ttf");
}
</style><body class="show_output">
<p>
 <a href="2d.text.draw.baseline.alphabetic.html" accesskey="p" title="[p] 2d.text.draw.baseline.alphabetic">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="2d.text.draw.baseline.hanging.html" accesskey="n" title="[n] 2d.text.draw.baseline.hanging">&gt;</a>
<h1><a href="index.2d.html">2d</a>.<a href="index.2d.text.html">text</a>.<a href="index.2d.text.draw.html">draw</a>.<a href="index.2d.text.draw.baseline.html">baseline</a>.ideographic</h1>
<p class="desc"></p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.2d.text.baseline.ideographic">2d.text.baseline.ideographic</a>

</ul>
</div>

<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.font = '50px CanvasTest';
deferTest();
setTimeout(wrapFunction(function () {
    ctx.fillStyle = '#f00';
    ctx.fillRect(0, 0, 100, 50);
    ctx.fillStyle = '#0f0';
    ctx.textBaseline = 'ideographic';
    ctx.fillText('CC', 0, 31.25);
    _assertPixelApprox(canvas, 5,5, 0,255,0,255, "5,5", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 95,5, 0,255,0,255, "95,5", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 25,25, 0,255,0,255, "25,25", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 75,25, 0,255,0,255, "75,25", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 5,45, 0,255,0,255, "5,45", "0,255,0,255", 2);
    _assertPixelApprox(canvas, 95,45, 0,255,0,255, "95,45", "0,255,0,255", 2);
}), 500);


});
</script>
