LayoutTests/platform/mac-snowleopard/canvas/philip/tests/2d.text.draw.baseline.ideographic-expected.txt

< [index] >
2d.text.draw.baseline.ideographic
References: 2d.text.baseline.ideographic
Actual output:
Expected output:

Failed assertion: got pixel [128,127,0,255] at (5,5), expected [0,255,0,255] +/- 2
Failed assertion: got pixel [128,127,0,255] at (95,5), expected [0,255,0,255] +/- 2
Failed assertion: got pixel [255,0,0,255] at (5,45), expected [0,255,0,255] +/- 2
Failed assertion: got pixel [255,0,0,255] at (95,45), expected [0,255,0,255] +/- 2
